Fix exception in callback_thread if the server stopped
New to Ruby so I don't know all the details, but stopping the server
for long enough causes .status to return true, which then causes
callback_thread to get an exception when it tries to access the status.

For now, I just take the quick'n'dirty route of checking the status's class.